DATABASE FINALLY CONNECTED!
Now on the index.jsp file we can make a connection that forms a SQL Query to request the products and list them in the main page. Only this feature is implemented, if you click on comprar, you will see that it doesn´t recognize the products and don´t appear
in the cart.

This last part will be available in future updates. Be ready!